Rename some test cases
* System culling is called only culling in the codebase.
* Simplified host dedup logic to just deduplication.
