+++ tests/test_api_put.py
@@ -6,7 +6,7 @@ from app.utils import HostWrapper
 from tests.test_api_utils import generate_uuid
 from tests.test_api_utils import HOST_URL
 from tests.test_api_utils import PreCreatedHostsBaseTestCase
-from tests.test_system_culling import QueryStalenessGetHostsBaseTestCase
+from tests.test_culling import QueryStalenessGetHostsBaseTestCase
 
 
 class FactsTestCase(PreCreatedHostsBaseTestCase):
