+++ tests/test_culling.py
